DF-8619 all keys works on win10
Former-commit-id: d7dd41fe838673795cfdd9f013e0780c5317ba11

+++ Sources/Internal/Debug/Replay.cpp
@@ -111,7 +111,7 @@ void Replay::RecordEvent(const UIEvent * ev)
 	Write(ev->phase);
 	Write(ev->controlState);
 	Write(ev->tapCount);
-	Write<uint16>(ev->keyChar);
+    Write<uint32>(ev->keyChar);
     Write(static_cast<uint32>(ev->device));
 }
